backend/requirements.txt
# Core dependencies
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
supabase==1.0.3

# Database
sqlalchemy==2.0.23
asyncpg==0.28.0

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Utilities
email-validator==2.1.0
python-multipart==0.0.6
